Modified delete these preferences to prompt the user to make sure they want to delete these preferences
Also handles no preferences saved under this key, by alerting the user and not calling deletePreferences.

+++ client/src/App.js
@@ -197,12 +197,19 @@ class KeyManagement extends Component{
   }
 
   async onDeletePreferences(){
-    // !!! Want to make an are you sure? confirmation
-    var success = await this.props.contract.methods.deletePreferences(hashKey(this.state.key)).send({from: this.props.address});
-    console.log(success); // !!! Message to user based on success (Preferences deleted!, or Preferences not found)
-    if(success){
-      this.setState({key: ''}, this.setStateKey); // Clears secret key input after preferences deleted
-      this.props.setPref('0000'); // Resets preference form back to default
+    if(window.confirm("Are you sure you want to delete these preferences?")){
+      try{
+        var retrPref = await this.props.contract.methods.getPreferences(this.props.address, hashKey(this.state.key)).call({from: this.props.address}); // Attempts to retrieve preferences for this address + key combo to see if there's anything to delete
+        var success = await this.props.contract.methods.deletePreferences(hashKey(this.state.key)).send({from: this.props.address});
+        console.log(success); // !!! Message to user based on success (Preferences deleted!, or Preferences not found)
+        if(success){
+          this.setState({key: ''}, this.setStateKey); // Clears secret key input after preferences deleted
+          this.props.setPref('0000'); // Resets preference form back to default
+      }
+    }
+      catch{
+        window.alert("No preferences found for this key, nothing to delete.");
+      }
     }
   }
